move routes to views.py
