+++ app.py
@@ -1,22 +1,7 @@
-from flask import Flask, render_template, request
+from flask import Flask
+from views import views
 app = Flask(__name__)
-
-
-@app.route('/')
-def dashboard():
-    return render_template('dashboard.html')
-
-@app.route('/signIn')
-def signIn():
-    return render_template('signIn.html')
-
-@app.route('/signUp')
-def signUp():
-    return render_template('signUp.html')
-
-@app.route('/signInCheck')
-def signInCheck():
-    return null
+app.register_blueprint(views, url_prefix="/")
 
 if __name__ == '__main__':
     app.run(debug=True, use_reloader=True)
